pcengines/apu2: run u-root command directly
The makefile now builds the u-root command when
make fetch
is invoked.

It runs the u-root command it builds, rather than
run u-root.go

Further, now that a set of bugs are fixed in cpud, we can now fire up
a console shell, and include lots of commands, which makes life a
LOT easier.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>