ampere/jade: Fix geturoot failling to download u-root/cpu
For go version 1.17, the 'go get' command is a combination of
downloading and installing a package. We have to use -d flag to skip
installing the package. For future go versions, the -d flag will be
enabled by default, so we have to use 'go install' to install the
package instead.

This patch also fixes the go/u-root version to avoid unexpected
compatibility issue between golang and u-root. Without specifying the
version, the go always gets the latest version.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>